minor housekeeping
