+++ src/cli-utils/index.ts
@@ -1,0 +1,2 @@
+// Export all CLI utility functions
+export * from "./install"
